Fixes bug with unnamed parameters
Unnamed parameters would go to the handler wrapper as if they were the
``context`` and ``nodelist`` arguments.  This basically boils down to
how the ``parse_bits`` utility minimizes the ``kwargs`` into positional
``args``, before we even get a chance to know about the context.

By zipping the ``args`` back to their original argument names,
everything becomes ``kwargs``, and later on when the partial is
consumed, the system doesn't become confused over thinking it already
had a positional value for ``context``.